also instantiate the slave and coreos-assembler in the template
This makes the set up instructions much easier and allows us finer
grained control over the resources in the future if we'd like. One
notable change here is `lookupPolicy.local: true`, which allows us to
directly reference the imagestream from the pod. This conveniently also
works around the registry DNS issues mentioned in #3.