pipeline: switch to scripted mode
Jenkins declarative vs scripted modes are a bit confusing. In the end, I
always find the declarative mode too limiting and verbose. It is not
uncommon for plugins to only offer some features through the more
powerful scripted mode rather than declarative.

Switching over allows us to make a cleaner Jenkinsfile without the yaml
definition of the containers baked in, thus reducing further the number
of nested languages we have to deal with.

We also avoid having to always nest commands inside `steps {}` and
`container() {}`. This considerably reduces the size of the file.