pipeline: add helper shwrap() function
Add a helper to execute shell commands in a consistent environment with
the traditional `set -xeuo pipefail` and always starting in the `/srv`
dir. I initially played with `workingDir: /srv`, but the Kubernetes
plugin tries to be smart and interprets it as a mountpoint to mount into
the container. Which we don't want here since we already have an `/srv`
mountpoint.

I also removed the `| tee` since I don't think it's helping. I think the
issue there is that the `rpm-ostree` running inside supermin thinks it's
got a tty, so we'd have to hack this directly in the init script.