pipeline: fix rsync for prod
Of course, just like `fileExists()` doesn't work well inside Kube
containers, neither does `readFile()`. Fix this by reverting back to
just reading and trimming in shell.

The larger issue here is that it's not easy to test the rsync bits
locally. Maybe we can fix this by replacing the simple-httpd stuff with
an artifact server deployment for local dev? Will look into it.