pipeline: autodetect if we're in prod/dev
As we add more stuff to the pipeline, it's going to get harder and
harder to transparently work for both the devel case and the prod case.
One example of this right now is rsync'ing: we use the detection of
whether the rsync key is mounted to know if we're in devel. But this is
not quite strong enough for the general case.

Let's just make an explicit check early on based on the `JENKINS_URL`
itself.