Bump Jenkins to v2.164.2
The switch to quay.io was a little premature here. The latest image is
for now only available from the previous location. Let's bump again to
v2.164.2, which is the latest LTS release.

This also fixes an issue with login which affected the previous
image[1], which somehow only manifested itself in prod.

[1] https://github.com/openshift/jenkins/issues/786