Make S3 bucket a parameter and optional
Don't require an S3 bucket to function. Instead, make it another
parameter. Also, make it optional so that one doesn't need S3 at all to
build (this was already somewhat the case through the `/.aws` mount,
though now we actually notice this and store builds in the PVC). Also
add a corresponding `--bucket` switch to `devel-up`.