pipeline: fully move to S3
Now that we have a prod build in S3, we can fully switch over to it.
Most of the changes are straightforward. One note of interest: the
"pruning" stage is now really just about pruning the local cache. For
now, we're not pruning from the bucket at all, pending a more defined
policy & mechanism.

Note we're still uploading the latest build to the artifact server to
make it easier for folks to download. Though soon we should replace that
with a frontend.