pipeline: stop using /srv as workdir
I think we should stop using the PVC as a workdir entirely and just
always build in the workspace. The core issue here is that (1) we want
to be able to have concurrent builds, and (2) a cosa workdir can't be
easily shared today. This also simplifies the devel vs prod logic quite
a bit since it had some funky conditionals around this.

Instead, let's only use the PVC to cache things. For now, we just go
with the naive approach of having separate directories per stream where
they can each store their `cache.qcow2`. Of course, this is highly
inefficient space-wise since many of the RPMs will be shared between
streams. We can refine this later on to build a smarter cache which
cosa/rpm-ostree can know about (e.g. by caching RPMs or keeping a
pkgcache repo).

For devel pipelines without S3 creds, we also keep the build result
itself in the PVC. Otherwise, there'd be no other way to access the
artifacts.