pipeline: add prod, devel, and mechanical pipelines
This adds pipeline buildconfigs for all the current streams (testing,
testing-devel, and bodhi-updates). Rather than having a separate
buildconfig for each stream, we just group them by prod, devel, and
mechanical since the triggering within a particular group is likely
going to be the same.

One note here is that I removed the GitHub webhook from the OpenShift
side, and instead made it part of the Jenkinsfile itself. This is
because we need webhooks to Jenkins anyway to trigger off of
`fedora-coreos-config`, so we might as well use the same mechanism for
`fedora-coreos-pipeline` as well.

Another note: I initially had the `prod` pipeline be solely manually
triggered, though changed that later on to be triggered on pushes to
`fedora-coreos-config`: I think it'd be cool if the only manual step
required for a release was simply to push a commit to `testing`. Anyway,
we can refine that later if needed (and all the other triggers as well).