pipeline: rename things to disambiguate
Right now, we have a "prod" pipeline, as well as "prod" streams. We have
"devel" pipelines, as well as "devel" streams. So e.g. one can have a
"devel prod" pipeline, which is of course not the same thing as the
"prod devel" pipeline.

Let's just rename things a bit here and establish a consistent
terminology to help disambiguate things:

- "production stream" --> refers to production streams (stable, testing,
  next)
- "development stream" --> refers to development streams (testing-devel,
  next-devel)
- "official pipeline" --> refers to the one true production pipeline
  which outputs official artifacts for all the streams
- "developer pipeline" --> refers to a developer's side pipeline