Make deploying easier, better, faster, stronger
So, this greatly simplifies the instructions for setting up. Basically
the `deploy` script (previously called `devel-up`) can now be used for
both the developer and official pipelines right after having
instantiated the Jenkins master.

Importantly, it can also *update* these resources. So we can use this
now directly even for the official pipeline after making changes to the
manifest, instead of having to edit by hand as before. We could even
hook this up to a job so it's automated.

By default, it will not create the stream wrapper pipelines in the
developer flow, unless `--all` is passed.