deploy: use `oc replace` instead of `oc apply`
I now understand better the difference between `oc apply` and `oc
replace`. The former accepts partial specs and is capable of updating
just parts of the objects, similar to a patch.  Whereas `oc replace`
will completely replace the object.

Here we want the latter. E.g. if we drop out a field from a list, we
also want it to drop out of the object on the next `./deploy`. Our
manifest should be canonical.