streams: use `--wait` so we wait and report errors if downstream fails
I actually wanted to have "blocking" semantics from the beginning, but
abandoned on that when switching away from the `build` step. And I
didn't think to try `oc start-build --wait`. There's `--follow` too, but
it doesn't work with Jenkins pipelines.

We can probably improve on this in the future by capturing the build
name, and printing a link to the Jenkins console for that build.