Jenkinsfile: start uploading AMIs
There are a few undesirable things in here, but it will do for now. E.g.
we're not replicating across multiple regions, nor are we uploading
separate vmdks.

There's a lot of stuff around this that the RHCOS team has internally.
We should look to upstreaming it into cosa (or replacing it with
existing functionality in mantle and extending as necessary), so that it
can be shared with FCOS.

For now, we're just patching things together so we at least have
`us-east-1` for preview.