Jenkinsfile: auto-release non-production stream builds
Being able to just launch the latest official `testing-devel` build in
AWS is super useful. For now at least, let's just "release" all the
non-production stream builds we do.

This way we also exercise the release code on the `testing-devel`
stream. I also think it'd be cool if we had Cincinnati also monitor
`testing-devel`, and just apply a dumb policy to allow us to exercise
automatic updates. This would be a prereq for that.