Jenkinsfile: only run Publish on new builds
Minor regression from #115. Now that the `Publish` stage is outside the
main cosa `podTemplate`, it always executes, even if no new builds were
created. This is because the no-op `return` only returns from that
scope.

Fix this by only assigning to `newBuildID` if it's actually new. Then
gate the `Publish` step on that.