Jenkinsfile: fix cache nuking logic
First, we were using `du` on the symlink, which always returned 0, so we
never actually nuked the cache. (This is a regression from when we
switched to using the PVC for caching the qcow2s only). Moving it to the
`Init` stage means we can directly use the real filename instead.

Second, we would (try to) nuke the cache after a build, but really we
want to handle this situation *before* a build. This code predates cosa
learning to regenerate the qcow2 at build time instead of only at `init`
time.

Finally, be slightly more conservative and knock the limit down to 7G.
The disk is 10G; so even if we just have 3G left, that should be plenty
for composing into (we should still probably implement some pkgcache
cleaning heuristics in rpm-ostree itself to make this less awkward
thought).