Jenkinsfile.mechanical: drop frequency to every 6h
Really, it's overkill to rebuild mechanical streams every 30m. The
reason is that we're inherently tied to Bodhi pushes, which runs on a
much slower cadence. Ideally, as the comment there indicates, we'd be
even smarter about change detection and only trigger when such events
happen.

(Note we still do immediately trigger a run on git pushes to a
mechanical branch in https://github.com/coreos/fedora-coreos-config).