Jenkinsfile: run Publish stage in main scope
So, this is somewhat a revert of #115 and #117. The other issue (other
than the one mentioned in #117) is that we now also `return` early if
kola tests failed. But we'd still try to run the `Publish` job.

In fact, there's a simpler way to do this. The `jnlp` container which
runs in the same pod as us also has `oc` pre-installed, and the
Kubernetes plugin makes it easy to run there using `container()`, so
let's just trigger it from there. And being under the main
`podTemplate()` scope again means that `return` really does mean
finishing the job early.