Jenkinsfile.*: consistently substitute COREOS_ASSEMBLER_MEMORY_REQUEST
We were only doing it in the main Jenkinsfile, but many of the other
ones use the same `pod.yaml` definition so we need to sub out that
variable there too.

In the future, I'd like to switch all jobs over to
[coreos-ci-lib](https://github.com/coreos/coreos-ci-lib) instead to make
this less error-prone and only actually mount the secrets each job
needs.