Jenkinsfile: set durability to PERFORMANCE_OPTIMIZED
By default Jenkins tries to be conservative and writes back lots of data
so that it can resume pipelines from a specific point if it gets
interrupted.

We don't care about that here. We want this to be a native functionality
in cosa (though we're not entirely there yet, see:
https://github.com/coreos/coreos-assembler/issues/668). And in the
future we want to split the pipeline into multiple jobs exactly to make
rerunning things easier.

For more information, see:
https://jenkins.io/doc/book/pipeline/scaling-pipeline/