Jenkinsfile: allow overriding cosa image to use
Add a `COREOS_ASSEMBLER_IMAGE` parameter that allows overriding the
default one to use. This is useful on its own, but it's necessary right
now for purposely building an FCOS image with older cosa.

In the future, I think we should take the cosa image we used when doing
a release and tag it. And e.g. keep it around for the last 5 releases
per stream? (The next level from this would be using lockfiles for cosa
too... would be cool, but way more overhead!)