pipeline: move Slack JCasC config to separate configmap
The configmap will contain more and more knobs. Rather than keep
expanding the template with more knobs, just split it out proper into
its own file. The downside is that users have to create it manually as a
separate step now. We could look at automating it again later.

This is needed for CoreOS CI which wants to share code with the
pipeline, but doesn't need Slack.