deploy: support sourcing multiple manifests
This patch makes preparations to have `deploy` be able to read and
instantiate from multiple manifests instead of just `pipeline.yaml`.

First, use a dict instead of a list to build the set of parameters for a
template. This will allow parameter filtering in the future, which is
needed so that we don't pass parameters to templates that don't support
them.

And second, make `process_template` return the list of resources
directly, instead of having other users later on needing to dig into the
`items` element for them.