Jenkinsfile: do `oc get pods -o wide` upfront
We're seeing occasional issues where pods can't find `/dev/kvm` even
though we have a `oci_kvm_hook: allowed` selector (and in fact
`oci-kvm-hook` is installed on the whole cluster).

Print out the pod list on startup so that we can figure out on which
node the pod was scheduled to help debugging the next time we hit this.