manifests/pod: stop using PVC cache for now
We've been hitting strange I/O related issues recently. We originally
traced it down to a specific node that was cordoned off, but now we're
hitting it on other nodes as well.

I think it might be related to the cache qcow2 being on an NFS and some
I/O patterns we do trigger some issues. Anyway, since the pipeline is
failing pretty bad, let's test that theory and swap it out for just an
`emptyDir`.

Also rename the mount to `cache` rather than `data` to make it clear we
don't actually need it.