jobs: add bump-lockfile
This job will implement lockfile bumping for testing-devel and
next-devel: https://github.com/coreos/fedora-coreos-tracker/issues/293.

The original plan for this functionality was to have it in config-bot:
https://github.com/coreos/fedora-coreos-releng-automation/pull/48

But in the end, I think it's more natural to have it as a Jenkins job
given that it does a lot of the same things as the pipeline/upstream CI
jobs. So that way it looks and feels just like another job that runs
cosa, and we get kola artifacts, we can re-use the shared library,
it's easily inspectable, we can hook it to Slack, etc...