update for GCP image build/release workflow
In our build/release workflow for GCP we will build and upload/create
a GCP image in the build pipeline. In the build pipeline we will
deprecate the image because we haven't decided to release it yet.
In the release pipeline we will undeprecate the image and deprecate
all other images in that image family using the ore gcloud promote-image
command.

Also attach the image to a license upon creation.