Jenkinsfile: attach nested virtualization license to GCP images
This will allow for launched instances to have access to /dev/kvm
by default. I am not aware of any known drawbacks to doing this as
I've been informed by GCP folks that if there is an environment that
doesn't have nested virt support or if there is a policy that disallows
it the instance simply won't have access to /dev/kvm. No other errors
would be presented to the user (like not being able to launch an
instance).