pipeline: archive liveiso on failure
This is a bit heavy-handed, but the pipeline is failing right now at the
`metal4k` `iso-offline-install` test and I can't reproduce it locally.
It's possible that there's a subtle mismatch, or it's somehow down to
the Jenkins environment. To rule that out, let's just archive the iso
for now if `testiso` failed.