bump-lockfile: add timeout and send Slack msg on failure
This job is key to our promotion model. We've had lapses in the past
where we didn't realize it was failing for a few days.

Let's send a Slack notification whenever the job fails to prevent this
(and yes, still hoping to eventually move to IRC notifications).

While we're here, also wrap it all in a timeout as good practice.