Convert kola test jobs to vanilla pipeline jobs
This is part of the push to move away from buildconfig-based jobs
towards the same model as what upstream CI does. For more details, see:

https://github.com/coreos/fedora-coreos-pipeline/pull/228

As a first victim, convert the kola tests since they're mostly
straight-forward. A huge part of this is re-using coreos-ci-lib the same
way upstream CI does. And this helps save us a lot of boilerplate (see
diffstat for this patch).

Converting to vanilla jobs also means we can now use the `build` step as
intended, instead of hacking around it with `oc start-build`.