jobs/sync-stream-metadata: sync one stream at a time
One obvious thing that I missed while working on #288 was that this job
doesn't actually have a `STREAM` parameter since it syncs all the
streams in one shot. So the fedmsg it emits will have `null` as the
stream name.

To make this stream-sensitive, let's change to doing `aws s3 sync`
separately for each stream with change detection for each. That way we
only emit fedmsgs for streams which actually had their metadata changed.

Patch best viewed with whitespace ignored.