plugins: add configuration-as-code-groovy
This is needed so that we can run some Groovy code via CASC dropins.
Obviously this goes against the declarative CASC premise a bit, but in
some cases you need that escape hatch. Here for example, we'll use this
to modify the pod concurrency cap on the *pre-existing* "openshift"
cloud that gets created (which it seems you can't do in pure CASC).