Jenkinsfile.production: trigger on GitHub pushes
So far, we've been triggering new production builds manually during the
release process.

In an effort to streamline that process, let's now switch to
automatically triggering a new Jenkins build whenever one of the
production branches is pushed to.

This does mean that we'll be triggering builds even when making changes
on the branch between releases (e.g. tweaking `manifest.yaml`). However,
this happens very rarely, and even then, it's totally fine if it
triggers a build. The pipeline is designed to allow multiple release
candidate builds.