bump-lockfile: strengthen no-op detection
Right now, we try to no-op early if we detect that the base lockfiles
didn't change. However, `--update-lockfile` will still touch the
lockfiles if just the repo timestamps changed (#232). This then causes
us to run `cosa build`, which no-ops, and then subsequent tests fail
because there is no e.g. qcow2 or ostree artifact that was built.

Fix this by detecting the no-op case just like the main pipeline does.

Closes: #232