bump-lockfile: compare lockfiles directly to determine no-op
The current `git diff` check we have here is good, but it's not entirely
foolproof because repo timestamps might have changed while the package
set did not.

Instead of using `git diff`, let's be more precise and specifically
compare the `packages` map in the files. That way we truly only ever do
any work (and push an update) if the package set changed.

Closes: #232