bump-lockfile: de-parallelize
Right now, the job tries to update both `testing-devel` and `next-devel`
in parallel, but this makes the output messy and hard to follow. It also
makes it harder to tell from the job status page which of the two
branches failed (and the BlueOcean UI isn't much help either).

Instead, let's make `bump-lockfile` focus on a single stream, and have a
driver job which simply triggers `bump-lockfile` once per stream so that
they're separate jobs. This mirrors how the main pipeline jobs
themselves are set up.

Patch best viewed with whitespace ignored.

Closes: #331