bump-lockfile: update metadata timestamp periodically
If we don't have package updates for the contentset of FCOS for a few
days, which is unusual but not unheard of (right now we haven't had
an updated RPM for 5 days), then bump the timestamp anyway. It's an
indicator of freshness for FCOS so even if no packages have changed
let's update it to bump the date in the FCOS version numbers.