release: automate waiting for import and pull test
In the FCOS release process, we have this step after the release job
runs where we want to wait until the OSTree commit was successfully
imported into the production repo and verify that it can be pulled and
that it's signed (though that last bit is also implicit with the ability
to even pull).

This requires human releasers to download the previous QEMU image, then
run `rpm-ostree upgrade` multiple times until it succeeds, and then look
at `rpm-ostree status`.

This patch enhances the release job to automate all that so that the
releaser just needs to verify the output. We use the exact same remote
config embedded in FCOS so e.g. commits must still be signed, mirrors
are used, etc...