Jenkinsfile: serialize access to builds.json
Since we are about to start forking off pipeline runs for
multiple architectures at the same time we need to add some
locking and higher level logic around builds.json. We only
want a single job accessing it at the same time and we only
want to add to it and never remove any build entries.

Create a new shared `bump_builds_json()` function that can
be used to implement this locking/logic.

We also add an early archive stage to the pipeline where we
buildupload the ostree and various metadata early on; and
bump the builds.json. This has the desired effect of "reserving"
our build ID earlier since after this point in the near future
we'll be forking off multiple other pipelines any of which
could fail (including x86_64). This means we're more like to
have build IDs from failed runs in our history, but that's OK.