Add multi-arch-pipeline job
This is the first attempt at adding a multi-arch pipeline. We'll
call the multi-arch-pipeline job from the main pipeline after QEMU
tests have passed.

This is currently mostly a copy/paste of the original pipeline. In
the future we want to factor out these bits and re-use them between
the main pipeline and the multi-arch-pipeline job.