Jenkinsfile.release: limit buildupload artifacts
In the x86_64 pipeline we early archive the ostree but the meta.json
that gets uploaded includes the qemu qcow. Right now the aarch64
pipeline is finishing before the x86_64 pipeline does and when we
go to buildupload it's complaining because the x86_64 qemu qcow
doesn't exist either locally or in s3.

```
Exception: builds/36.20210812.91.0/x86_64/fedora-coreos-36.20210812.91.0-qemu.x86_64.qcow2 not found locally or in the s3 destination!
```

Since we shouldn't be uploading image artifacts in the release step
anyway, let's just pass --artifact=ostree, which will essentially
give us what we want. It won't re-upload the ostree because it already
exists in s3, and it won't try any other artifact, including qemu qcow.