release: Copy oscontainer to quay.io
This is a bit hacky but should work.  My initial goal here
is just automated uploads of our builds, so that we can
start getting a better feel for managing "ostree-in-container"
releases.

We should sync to `quay.io/coreos` but that needs someone to set that up.

Also, I'd like to make the destination configurable in the same
way as the S3 bucket, proposal PR in https://github.com/coreos/fedora-coreos-config/pull/1175

Closes: https://github.com/coreos/fedora-coreos-pipeline/issues/359