multi-arch-pipeline: allow for failing as soon as tests fail
Moving the bulk of the commands into `commands:`, leaving the `tar`
in `post_commands:` and setting `post_always` to true will allow the
tarball to get created always. This will allow for us to fail as
soon as a kola run fails. It also means we don't have to rely on
reading the report.json, which happens to not get created for testiso
runs, and was never going to work.