multi-arch-pipeline: fix tar invocation and reorder post_commands
The previous iteration `ls` was actually returning a bad exit code
if some directories didn't exist, which meant the first list item
in `post_commands` would fail and the compress wouldn't happen. Let's
move the compress to the first step and also change the tar command
to use `|| true` so it won't "fail" if we add any more post commands
in the future.

Note that we can get rid of the `compress` in the future when we have
https://github.com/coreos/coreos-assembler/issues/2389