kola-openstack: switch to v3-starter-4 openstack flavor
Despite extending the timeout to 10 minutes we're still hitting
timeouts waiting for nodes to come up in vexxhost. Reaching out
to the CEO he suggested trying the v3 instances (newer infra) to
see if that helped.

If this doesn't work we can use the equivalent of `--boot-from-volume`
to get EBS like instances (shouldn't take time to copy image to local
compute nodes).