Use PERFORMANCE_OPTIMIZED for all jobs
Otherwise, the default is to have Jenkins retain the state of jobs so
that it can resume them in the event of a loss of power/restart. But we
don't really need that feature because our jobs are designed to be
idempotent and e.g. we need to be able to handle restarting from
scratch.

So disable it so we can win the performance gains.

Prompted by recently-added (#365) code which breaks the job because
Jenkins isn't capable of serializing all the state to disk.