kola-{aws,gcp,openstack}: use a specific config git commit for tests
This will allow us to only run tests that were for the specific config
repo version that we built against. This prevents issues like what we saw
recently where tests got kicked off later and the config repo had merged
a new change that had a test. The test failed because the test was ran
but the corresponding needed change was not in the built FCOS version.